+++ x-pack/plugins/search_playground/__mocks__/search_playground_mock.ts
@@ -14,6 +14,7 @@ const createStartMock = (): Start => {
     PlaygroundProvider: jest.fn(),
     PlaygroundToolbar: jest.fn(),
     Playground: jest.fn(),
+    PlaygroundHeaderDocs: jest.fn(),
   };
 
   return startContract;
