+++ x-pack/plugins/search_playground/kibana.jsonc
@@ -14,6 +14,7 @@
       "actions",
       "encryptedSavedObjects",
       "navigation",
+      "share",
       "security",
       "stackConnectors",
       "triggersActionsUi",
