+++ x-pack/plugins/search_playground/public/application.tsx
@@ -7,53 +7,26 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { CoreStart, AppMountParameters } from '@kbn/core/public';
+import { CoreStart } from '@kbn/core/public';
 import { KibanaThemeProvider } from '@kbn/react-kibana-context-theme';
 import { KibanaContextProvider } from '@kbn/kibana-react-plugin/public';
 import { I18nProvider } from '@kbn/i18n-react';
-import { BrowserRouter as Router } from '@kbn/shared-ux-router';
-import { i18n } from '@kbn/i18n';
-import { KibanaPageTemplate } from '@kbn/shared-ux-page-kibana-template';
-import { PLUGIN_ID } from '../common';
+import { Router } from '@kbn/shared-ux-router';
 import { AppPluginStartDependencies } from './types';
-import { App } from './components/app';
-import { PlaygroundProvider } from './providers/playground_provider';
 
-export const renderApp = (
+export const renderApp = async (
   core: CoreStart,
   services: AppPluginStartDependencies,
-  { appBasePath, element }: AppMountParameters
+  element: HTMLElement
 ) => {
-  const navigation = services.navigation;
+  const { PlaygroundRouter } = await import('./playground_router');
 
   ReactDOM.render(
     <KibanaThemeProvider theme={core.theme}>
       <KibanaContextProvider services={{ ...core, ...services }}>
         <I18nProvider>
-          <Router basename={appBasePath}>
-            <navigation.ui.TopNavMenu appName={PLUGIN_ID} />
-            <PlaygroundProvider
-              defaultValues={{
-                indices: [],
-              }}
-            >
-              <KibanaPageTemplate
-                pageChrome={[
-                  i18n.translate('xpack.searchPlayground.breadcrumb', {
-                    defaultMessage: 'Playground',
-                  }),
-                ]}
-                pageHeader={{
-                  pageTitle: i18n.translate('xpack.searchPlayground.pageTitle', {
-                    defaultMessage: 'Playground',
-                  }),
-                }}
-                bottomBorder="extended"
-                restrictWidth={false}
-              >
-                <App />
-              </KibanaPageTemplate>
-            </PlaygroundProvider>
+          <Router history={services.history}>
+            <PlaygroundRouter />
           </Router>
         </I18nProvider>
       </KibanaContextProvider>
