+++ x-pack/plugins/search_playground/public/routes.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const ROOT_PATH = '/';
+export const SEARCH_PLAYGROUND_CHAT_PATH = `${ROOT_PATH}chat`;
