+++ x-pack/plugins/serverless_search/tsconfig.json
@@ -47,5 +47,6 @@
     "@kbn/discover-plugin",
     "@kbn/search-connectors-plugin",
     "@kbn/index-management",
+    "@kbn/search-playground",
   ]
 }
